+++ NEWS.md
@@ -1,3 +1,25 @@
+# ospsuite.utils 1.2.0
+
+MAJOR CHANGES
+
+* Removes alias `hasUniqueValues()`.
+
+* All messages used in `ospsuite` package are now in `ospsuit.utils` (exported
+  list `messages`).
+
+BUG FIXES
+
+* Fixes a regression in `isIncluded()` in previous release that inadvertently
+  removed support for compound types (#63).
+
+NEW FUNCTIONS
+
+* Adds `isEmpty()` and `validateIsNotEmpty()` functions to validate that objects
+  aren't empty (#58; thanks to @pchelle).
+
+* Adds `getOSPSuiteUtilsSetting()` function to get global settings
+  (see enum `ospsuiteUtilsSettingNames` for supported settings).
+
 # ospsuite.utils 1.1.0
 
 MAJOR CHANGES
@@ -10,11 +32,13 @@ MINOR CHANGES
 
 * Improvements to documentation.
 
-* `validateIsCharacter()` is added as an alias for `validateIsString()` function.
+* `validateIsCharacter()` is added as an alias for `validateIsString()`
+  function.
 
 * `getEnumKey()` is added as an alias for `enumGetKey()` function.
 
-* `hasOnlyDistinctValues()` is added as an alias for `hasUniqueValues()` function.
+* `hasOnlyDistinctValues()` is added as an alias for `hasUniqueValues()`
+  function.
 
 BUG FIXES